fixes to cpuaffinity
